drop php7.2 and mysql for openrc xen-guest added

+++ autobuild.sh
@@ -10,45 +10,48 @@ if [ "$EUID" -ne 0 ]; then
 fi
 
 PROFILE=$1
-
-if [ ! -x profile/$PROFILE/stage3-url.sh ]; then
-  echo "Need profile/$PROFILE/stage3-url.sh which shows stage3 url to build gentoo corresponds to the profile."
-  exit 1
-fi
-
-STAGE3_URL=$(sh profile/$PROFILE/stage3-url.sh)
-if [ -z "$STAGE3_URL" ]; then
-	echo "Stage3 URL could not be determined."
-	exit 1
-fi
-echo "Stage3 tarball is $STAGE3_URL ."
-STAGE3_HASH=$(echo -n "$STAGE3_URL"|md5sum|cut -f 1 -d " ")
-STAGE3_FILE=cache/download/$STAGE3_HASH
-
-if [ ! -f "$STAGE3_FILE" ]; then
-	TEMPFILE=cache/download/download-$$.tmp
-	mkdir -p cache/download
-	if wget -O $TEMPFILE $STAGE3_URL; then
-		mv $TEMPFILE $STAGE3_FILE
-	else
-		exit 1
-	fi
-fi
-
 GENTOO_DIR=gentoo/$PROFILE
 
-[ -f "${GENTOO_DIR}/done" ] && DONE=$(cat "${GENTOO_DIR}/done")
+if [ -z "$SKIP_INIT" ]; then
 
-if [ "$DONE" != "$STAGE3_HASH" ]; then
-  echo "Initializing $GENTOO_DIR ..."
-  $SUDO rm -rf "$GENTOO_DIR"
-  $SUDO mkdir -p $GENTOO_DIR
-  echo "Extracting stage3..."
-  $SUDO tar xpf $STAGE3_FILE -C $GENTOO_DIR || exit 1
-  $SUDO mkdir -p $GENTOO_DIR/var/db/repos/gentoo || exit 1
-  $SUDO sh -c "echo 'GENTOO_MIRRORS=\"http://ftp.iij.ad.jp/pub/linux/gentoo/\"' >> $GENTOO_DIR/etc/portage/make.conf"
-  $SUDO sh -c "echo 'FEATURES=\"-sandbox -usersandbox -network-sandbox\"' >> $GENTOO_DIR/etc/portage/make.conf"
-fi
+  if [ ! -x profile/$PROFILE/stage3-url.sh ]; then
+    echo "Need profile/$PROFILE/stage3-url.sh which shows stage3 url to build gentoo corresponds to the profile."
+    exit 1
+  fi
+
+  STAGE3_URL=$(sh profile/$PROFILE/stage3-url.sh)
+  if [ -z "$STAGE3_URL" ]; then
+  	echo "Stage3 URL could not be determined."
+  	exit 1
+  fi
+  echo "Stage3 tarball is $STAGE3_URL ."
+  STAGE3_HASH=$(echo -n "$STAGE3_URL"|md5sum|cut -f 1 -d " ")
+  STAGE3_FILE=cache/download/$STAGE3_HASH
+
+  if [ ! -f "$STAGE3_FILE" ]; then
+  	TEMPFILE=cache/download/download-$$.tmp
+  	mkdir -p cache/download
+  	if wget -O $TEMPFILE $STAGE3_URL; then
+  		mv $TEMPFILE $STAGE3_FILE
+  	else
+  		exit 1
+  	fi
+  fi
+
+  [ -f "${GENTOO_DIR}/done" ] && DONE=$(cat "${GENTOO_DIR}/done")
+
+  if [ "$DONE" != "$STAGE3_HASH" ]; then
+    echo "Initializing $GENTOO_DIR ..."
+    $SUDO rm -rf "$GENTOO_DIR"
+    $SUDO mkdir -p $GENTOO_DIR
+    echo "Extracting stage3..."
+    $SUDO tar xpf $STAGE3_FILE -C $GENTOO_DIR || exit 1
+    $SUDO mkdir -p $GENTOO_DIR/var/db/repos/gentoo || exit 1
+    $SUDO sh -c "echo 'GENTOO_MIRRORS=\"http://ftp.iij.ad.jp/pub/linux/gentoo/\"' >> $GENTOO_DIR/etc/portage/make.conf"
+    $SUDO sh -c "echo 'FEATURES=\"-sandbox -usersandbox -network-sandbox\"' >> $GENTOO_DIR/etc/portage/make.conf"
+  fi
+
+fi # SKIP_INIT
 
 $SUDO cp /etc/resolv.conf $GENTOO_DIR/etc/resolv.conf
 #$SUDO sh -c "echo -e 'nameserver 8.8.8.8\nnameserver 8.8.4.4' > $GENTOO_DIR/etc/resolv.conf"
